+++ src/containers/SendMessageForm.tsx
@@ -33,7 +33,7 @@ import { Modal } from "../components/Common/modal";
 import { Button } from "../components/Common/Button";
 import { MAX_PAYLOAD_SIZE } from "../config/constants";
 import { prepareFileForUpload } from "../utils/upload-file-handler";
-import { useIsMobile } from "../utils/useIsMobile";
+import { useIsMobile } from "../hooks/useIsMobile";
 import { parseImageFileJson } from "../utils/parse-image-file";
 
 type SendMessageFormProps = {
